Add AddSample and RemoveAllSample

+++ gdataset.c
@@ -1186,17 +1186,16 @@ bool GDSVecFloatSaveCategory(
 
 // Run the prediction by the NeuraNet 'nn' on each sample of the category
 // 'iCat' of the GDataSet 'that'. The index of columns in the samples
-// for inputs and outputs are given by 'inputs' and 'outputs'.
-// input values in [-1,1] and output values in [-1,1]
+// for inputs and outputs are given by 'iInputs' and 'iOutputs'.
 // Stop the prediction of samples when the result can't get better
-// than 'threhsold'
+// than 'threshold'
 // Return the value of the NeuraNet on the predicted samples, defined
 // as sum_samples(||output_sample-output_neuranet||)/nb_sample
 // Higher is better, 0.0 is best value
 float GDataSetVecFloatEvaluateNN(
   const GDataSetVecFloat* const that, 
   const NeuraNet* const nn, 
-  const long iCat, 
+  const int iCat, 
   const VecShort* const iInputs,
   const VecShort* const iOutputs,
   const float threshold) {
